requirements.txt
langgraph>=0.2.0
typing_extensions>=4.8.0
python-dotenv>=1.0.1
openai>=1.3.0
nbformat>=5.10.0
tiktoken>=0.7.0; platform_system != "Windows" or platform_machine != "ARM64"

